requirements.txt
asgiref==3.2.3
astroid==2.4.2
autopep8==1.4.4
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.42.0
django-bootstrap4==1.1.1
django-cors-headers==3.3.0
django-countries==6.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==1.10.1
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-paypal==1.0.0
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
newsapi-python==0.2.6
oauthlib==3.0.1
pep8==1.7.1
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.5.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.2.0
six==1.15.0
soupsieve==2.0
sqlparse==0.3.0
stripe==2.27.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.12.1
